% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{html_to_csv_prompt}
\alias{html_to_csv_prompt}
\title{Input: message content template for html_to_csv()}
\format{
## `html_to_csv_prompt`
A data frame with 2 rows and 2 columns representing the messages object in API calls.
}
\source{
Contributed by umatter
}
\usage{
html_to_csv_prompt
}
\description{
This message content serves as user input for the html_to_csv function.
}
\keyword{datasets}
