% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_csv.R
\name{xml_to_csv}
\alias{xml_to_csv}
\title{Convert XML input to CSV}
\usage{
xml_to_csv(file)
}
\arguments{
\item{file}{A character string representing the path to the file with XML input}
}
\value{
If the input is a character string, the function returns the output
as a character string and prints it to the console. Otherwise, the function returns
the output as a data frame and generates a CSV file with the same name as the input file
but with a .csv extension.
}
\description{
This function takes a file with XML input and processes it to generate
corresponding output in CSV format. If the input is a character string,
the output will be printed to the console. Otherwise, a CSV file will be
generated with the same name as the input file but with a .csv extension.
}
\examples{
\dontrun{
xml_string <- '<?xml version="1.0"?>
<employees>
 <employee>
   <firstName>John</firstName>
   <lastName>Doe</lastName>
 </employee>
 <employee>
   <firstName>Anna</firstName>
   <lastName>Smith</lastName>
 </employee>
 <employee>
   <firstName>Peter</firstName>
   <lastName>Jones</lastName>
 </employee>
</employees>'
xml_to_csv(xml_string)
}
}
