% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_csv.R
\name{html_to_csv}
\alias{html_to_csv}
\title{Convert HTML input to CSV}
\usage{
html_to_csv(file)
}
\arguments{
\item{file}{A character string representing the path to the file with HTML input}
}
\value{
If the input is a character string, the function returns the output
as a character string and prints it to the console. Otherwise, the function returns
the output as a data frame and generates a CSV file with the same name as the input file
but with a .csv extension.
}
\description{
This function takes a file with HTML input and processes it to generate
corresponding output in CSV format. If the input is a character string,
the output will be printed to the console. Otherwise, a CSV file will be
generated with the same name as the input file but with a .csv extension.
}
\examples{
\dontrun{
html_string <- '<table>
 <tr>
   <th>firstName</th>
   <th>lastName</th>
 </tr>
 <tr>
   <td>John</td>
   <td>Doe</td>
 </tr>
 <tr>
   <td>Anna</td>
   <td>Smith</td>
 </tr>
 <tr>
   <td>Peter</td>
   <td>Jones</td>
 </tr>
</table>'
html_to_csv(html_string)
}
}
